+++ sdk/servicebus/service-bus/src/receivers/receiver.ts
@@ -9,7 +9,7 @@ import {
   SubscribeOptions,
   InternalMessageHandlers
 } from "../models";
-import { OperationOptionsBase } from "../modelsToBeSharedWithEventHubs";
+import { OperationOptionsBase, trace } from "../modelsToBeSharedWithEventHubs";
 import { ServiceBusReceivedMessage } from "..";
 import { ConnectionContext } from "../connectionContext";
 import {
@@ -23,12 +23,12 @@ import { OnError, OnMessage, ReceiveOptions } from "../core/messageReceiver";
 import { CreateStreamingReceiverOptions, StreamingReceiver } from "../core/streamingReceiver";
 import { BatchingReceiver } from "../core/batchingReceiver";
 import { assertValidMessageHandlers, getMessageIterator, wrapProcessErrorHandler } from "./shared";
-import { convertToInternalReceiveMode } from "../constructorHelpers";
 import Long from "long";
 import { ServiceBusReceivedMessageWithLock, ServiceBusMessageImpl } from "../serviceBusMessage";
 import { Constants, RetryConfig, RetryOperationType, RetryOptions, retry } from "@azure/core-amqp";
 import "@azure/core-asynciterator-polyfill";
 import { LockRenewer } from "../core/autoLockRenewer";
+import { createProcessingSpan } from "../diagnostics/instrumentServiceBusMessage";
 import { receiverLogger as logger } from "../log";
 
 /**
@@ -158,6 +158,8 @@ export class ServiceBusReceiverImpl<
   private _streamingReceiver?: StreamingReceiver;
   private _lockRenewer: LockRenewer | undefined;
 
+  private _createProcessingSpan: typeof createProcessingSpan;
+
   private get logPrefix() {
     return `[${this._context.connectionId}|receiver:${this.entityPath}]`;
   }
@@ -179,6 +181,7 @@ export class ServiceBusReceiverImpl<
       maxAutoRenewLockDurationInMs,
       receiveMode
     );
+    this._createProcessingSpan = createProcessingSpan;
   }
 
   private _throwIfAlreadyReceiving(): void {
@@ -229,7 +232,7 @@ export class ServiceBusReceiverImpl<
     onInitialize: () => Promise<void>,
     onMessage: OnMessage,
     onError: OnError,
-    options?: SubscribeOptions
+    options: SubscribeOptions
   ): void {
     this._throwIfReceiverOrConnectionClosed();
     this._throwIfAlreadyReceiving();
@@ -245,7 +248,7 @@ export class ServiceBusReceiverImpl<
 
     this._createStreamingReceiver(this._context, this.entityPath, {
       ...options,
-      receiveMode: convertToInternalReceiveMode(this.receiveMode),
+      receiveMode: this.receiveMode,
       retryOptions: this._retryOptions,
       cachedStreamingReceiver: this._streamingReceiver,
       lockRenewer: this._lockRenewer
@@ -263,7 +266,9 @@ export class ServiceBusReceiverImpl<
         }
 
         if (!this.isClosed) {
-          sReceiver.subscribe(onMessage, onError);
+          sReceiver.subscribe(async (message) => {
+            await onMessage(message);
+          }, onError);
         } else {
           await sReceiver.close();
         }
@@ -297,7 +302,7 @@ export class ServiceBusReceiverImpl<
       if (!this._batchingReceiver || !this._context.messageReceivers[this._batchingReceiver.name]) {
         const options: ReceiveOptions = {
           maxConcurrentCalls: 0,
-          receiveMode: convertToInternalReceiveMode(this.receiveMode),
+          receiveMode: this.receiveMode,
           lockRenewer: this._lockRenewer
         };
         this._batchingReceiver = this._createBatchingReceiver(
@@ -306,12 +311,14 @@ export class ServiceBusReceiverImpl<
           options
         );
       }
+
       const receivedMessages = await this._batchingReceiver.receive(
         maxMessageCount,
         options?.maxWaitTimeInMs ?? Constants.defaultOperationTimeoutInMs,
         defaultMaxTimeAfterFirstMessageForBatchingMs,
-        options?.abortSignal
+        options ?? {}
       );
+
       return (receivedMessages as unknown) as ReceivedMessageT[];
     };
     const config: RetryConfig<ReceivedMessageT[]> = {
@@ -351,17 +358,12 @@ export class ServiceBusReceiverImpl<
     const receiveDeferredMessagesOperationPromise = async () => {
       const deferredMessages = await this._context
         .getManagementClient(this.entityPath)
-        .receiveDeferredMessages(
-          deferredSequenceNumbers,
-          convertToInternalReceiveMode(this.receiveMode),
-          undefined,
-          {
-            ...options,
-            associatedLinkName: this._getAssociatedReceiverName(),
-            requestName: "receiveDeferredMessages",
-            timeoutInMs: this._retryOptions.timeoutInMs
-          }
-        );
+        .receiveDeferredMessages(deferredSequenceNumbers, this.receiveMode, undefined, {
+          ...options,
+          associatedLinkName: this._getAssociatedReceiverName(),
+          requestName: "receiveDeferredMessages",
+          timeoutInMs: this._retryOptions.timeoutInMs
+        });
       return (deferredMessages as any) as ReceivedMessageT[];
     };
     const config: RetryConfig<ReceivedMessageT[]> = {
@@ -426,6 +428,7 @@ export class ServiceBusReceiverImpl<
     close(): Promise<void>;
   } {
     assertValidMessageHandlers(handlers);
+    options = options ?? {};
 
     const processError = wrapProcessErrorHandler(handlers);
 
@@ -440,7 +443,8 @@ export class ServiceBusReceiverImpl<
         }
       },
       async (message: ServiceBusMessageImpl) => {
-        return handlers.processMessage((message as any) as ReceivedMessageT);
+        const span = this._createProcessingSpan(message, this, this._context.config, options);
+        return trace(() => handlers.processMessage((message as any) as ReceivedMessageT), span);
       },
       processError,
       options