+++ sdk/storage/storage-blob/CHANGELOG.md
@@ -1,7 +1,14 @@
 # Release History
 
-## 12.2.2 (Unreleased)
+## 12.3.0-beta.2 (Unreleased)
 
+
+## 12.3.0-beta.1 (2020-10-13)
+
+- Updated Azure Storage Service API version to 2020-02-10.
+- Added support for Blob Last Access Time tracking.
+- Added support for Blob Query Arrow output format.
+- Added support for Container Soft Delete.
 
 ## 12.2.1 (2020-09-17)